% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{msa_palettes}
\alias{msa_palettes}
\alias{msa_palette_DNA}
\alias{msa_palette_RNA}
\alias{msa_palette_AA}
\title{Default colour palettes for nucleic acid and amino acid alignments}
\format{
An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{character} of length 21.
}
\usage{
msa_palette_DNA

msa_palette_RNA

msa_palette_AA
}
\value{
A named character vector of hex colour strings.
}
\description{
Named character vectors that map each letter of an alphabet to a hex
colour, suitable for passing as \verb{letter.colours =} to
\code{\link[=msaHeatmap]{msaHeatmap()}} when \code{column = "Letter"}.
}
\details{
\code{msa_palette_DNA} covers \verb{A C G T N}. \code{msa_palette_RNA} is the same
palette with \code{T} replaced by \code{U}. \code{msa_palette_AA} covers the 20
standard amino acids plus \code{X} (unknown) and follows a Clustal-X-style
physico-chemical colour scheme.
}
\examples{
msa_palette_DNA
msa_palette_AA

aln <- c(seq1 = "ACGTACGT", seq2 = "ACGTTCGT")
msaHeatmap(msa2DF(aln, reference = "seq1"),
  column = "Letter", letter.colours = msa_palette_DNA)

}
\keyword{datasets}
