% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaviz.R
\docType{package}
\name{msaviz-package}
\alias{msaviz}
\alias{msaviz-package}
\title{msaviz: Multiple sequence alignment visualization}
\description{
A small set of functions for turning a multiple sequence alignment into a
tidy data frame and then visualizing it, either as a ggplot2 heatmap or as
a per-sequence lollipop chart, with a few companion plots that compose
cleanly alongside it.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bjmt/msaviz}
  \item Report bugs at \url{https://github.com/bjmt/msaviz/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Jean-Marie Tremblay \email{benjmtremblay@gmail.com} (\href{https://orcid.org/0000-0002-7441-2951}{ORCID})

}
\keyword{internal}
