% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msaviz.R
\docType{package}
\name{msaviz-package}
\alias{msaviz}
\alias{msaviz-package}
\title{msaviz: Multiple sequence alignment visualization}
\description{
Provides a handful of useful functions for simplifying visualizations of
multiple sequence alignments using heatmaps and lollipop charts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bjmt/msaviz}
  \item Report bugs at \url{https://github.com/bjmt/msaviz/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Jean-Marie Tremblay \email{benjmtremblay@gmail.com} (\href{https://orcid.org/0000-0002-7441-2951}{ORCID})

}
\keyword{internal}
